pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "soundfile>=0.12.1",
    "numpy>=2.1.2",
    "pydub>=0.25.1",
    "azure-cognitiveservices-speech>=1.41.1",
    "aiohttp>=3.10.10",
    "websockets>=13.1",
    "python-dotenv>=1.0.1",
    "deepgram-sdk==2.12.0",
]
